+++ MST-game-toio-30/Assets/Scripts/DemoSync/DemoGUIManager.cs
@@ -46,9 +46,9 @@ public class DemoGUIManager : MonoBehaviour
                 GUILayout.Label("Collision Vibration Intensity: ");
                 manager.m_puppetCollisionVibrationIntensity = (int)GUILayout.HorizontalSlider((float)manager.m_puppetCollisionVibrationIntensity, 0.0f, 100.0f);
                 GUILayout.Label("Collision Tolerance: ");
-                manager.m_puppetCollisionTolerance = GUILayout.HorizontalSlider(manager.m_puppetCollisionTolerance, 0.0f, 100.0f);
+                manager.m_puppetCollisionTolerance = (int)GUILayout.HorizontalSlider((float)manager.m_puppetCollisionTolerance, 0.0f, 100.0f);
                 GUILayout.Label("Puppet Speed: ");
-                manager.m_puppetCollisionTolerance = GUILayout.HorizontalSlider(manager.m_puppetSpeed, 0.0f, 100.0f);
+                manager.m_puppetCollisionTolerance = (int)GUILayout.HorizontalSlider((float)manager.m_puppetSpeed, 0.0f, 100.0f);
             }
 
             GUILayout.EndArea();
